requirements.txt
django==5.0.3
djangorestframework==3.14.0
black==24.1.1
psycopg2-binary==2.9.9
ruff==0.1.15